+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * `create_report`: bump bioconductor version to 3.21 in order to accommodate R version 4.5 (PR #52).
 
+## Bug fixes
+
+* `reporting`: updated default colour mapping (PR #50).
+
 # htrnaseq v0.7.0
 
 ## Breaking changes